Move `AuthorOverlay` into `quill-util`.
Its former module `remote-authors` was a single-class module and was
mutually-dependent with `quill-util`.
