+++ local-modules/quill-util/package.json
@@ -9,7 +9,7 @@
     "doc-client": "local",
     "doc-common": "local",
     "hooks-client": "local",
-    "remote-authors": "local",
+    "typecheck": "local",
     "util-common": "local"
   }
 }
