--- local-modules/remote-authors/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "remote-authors",
-  "version": "1.0.0",
-  "main": "main.js"
-}
